requirements.txt
python-dotenv~=1.0.0
numpy~=1.26.2
requests~=2.31.0
z3-solver
networkx~=3.2.1